First task to redesign a website landing page
So my first task with this company was to redesign a website landing page. I tried looking for several websites on which i could work upon, but couldn't find one. So i got the idea to look upon the official website of the company i'm working for. When i went to that website, i was shocked by the effortless interface of the landing page. the original orange and white color combination felt soo unsettling. In my opinion, white color should be avoided from every website or applications as its very uncomfortable for the eyes of the user, hence i always prefer dark background for the websites and applications. For redesigning their website, i took the reference from a website "VERCEL.COM". I think its dark background feels soo aesthestic along with providing comfort to the eyes. I used an image of milky way for the background of the webpage, which resembles the infinite capability of the company. For fonts i used Italic since it looks creative along with professional. I also removed the unnecessary image of the employees  which was taking unnecessary space for the mobile users. And for the final touch, i changed the taglines of the webpage. Ya, i also arranged the buttons like products, resources, docs etc.