INTERNSHIP TASK 2
So my next task was to create a webpage which could adapt the screen size of different display by the help of HTML, JAVASCRIPT and CSS. In order to complete this task, i used a website called "CODEPEN.COM', because this website is very user friendly and easy to use. For codes, i took the help of chatgpt. i tried to recreate the webpage of the company im working under. Here is the direct link for the webpage:

https://codepen.io/Soumay-the-typescripter/pen/myJvWWo

From the link above, user can easily understand my work. 
thank you.