INTERNSHIP TASK 3
This is the 3rd internship task, in which i had to redesign the ui/ux of a random mobile application. After searching for a long time, i made my mind to redesign amazon shopping app. In original application, i observed that the original ui of the application is very hectic. everything feels soo messed up, which is very less user-friendly. what i did was creating a more user-friendly ui for the application. Also when buying something from this application, i noticed that the descriptions are also overdone, so i tried solving that issue as well.