INTERNSHIP TASK 4
So in this finl task, i had to create a fully fuctioning vr application. So i made a VR gaming application with the help of figma. Though i didn't had the premium version of a plugin called DRAFTXR, i tried making it as much better as possible. And with this task, my internship is finally over. I'll provide the link for my prototype that i made for VR application.

https://app.draftxr.com/vr/GVAwci